+++ src/multi_bot_agentic/tools/__init__.py
@@ -49,6 +49,7 @@ from multi_bot_agentic.tools.json_query import JsonQueryTool
 from multi_bot_agentic.tools.json_unflatten import JsonUnflattenTool
 from multi_bot_agentic.tools.jsonl_parse import JsonlParseTool
 from multi_bot_agentic.tools.jwt_decode import JwtDecodeTool
+from multi_bot_agentic.tools.levenshtein import LevenshteinTool
 from multi_bot_agentic.tools.line_number import LineNumberTool
 from multi_bot_agentic.tools.markdown_table import MarkdownTableTool
 from multi_bot_agentic.tools.markdown_toc import MarkdownTocTool
@@ -148,6 +149,7 @@ __all__ = [
     "JsonUnflattenTool",
     "JsonlParseTool",
     "JwtDecodeTool",
+    "LevenshteinTool",
     "LineNumberTool",
     "MarkdownTableTool",
     "MarkdownTocTool",